Update travel-guide-1-book-travel.md
*If you don't yet have a travel card and book through AdTrav, flights or rail will still be charged directly to the GSA, however, hotels must be charged against a personal credit or debit card.*

I'm not sure if this is better or not but I'm going for making it more clear that the traveler will have to use their own credit/debit card then get reimbursed.  As a first time traveler I might find this a little unsettling.  

So, with this note, should we also say something after like: Hotel expenses are still reimbursable so long as...something about being compliant with TTS travel policy or under the gov's standard rate for the area?